+++ src/internal/compilerSupport/v1.ts
@@ -1,4 +1,4 @@
-export { createClient, FluenceClient } from '../../FluenceClient';
+export { FluencePeer } from '../FluencePeer';
 export { ResultCodes } from '../../internal/CallServiceHandler';
 export { RequestFlow } from '../../internal/RequestFlow';
 export { RequestFlowBuilder } from '../../internal/RequestFlowBuilder';
