+++ src/internal/utils.ts
@@ -1,0 +1,89 @@
+import { AirInterpreter, LogLevel as AvmLogLevel } from '@fluencelabs/avm';
+import log from 'loglevel';
+import { FluencePeer } from './FluencePeer';
+import { RequestFlowBuilder } from './RequestFlowBuilder';
+
+export const createInterpreter = (handler, peerId): Promise<AirInterpreter> => {
+    let logLevel: AvmLogLevel = 'off';
+    switch (log.getLevel()) {
+        case 0: // 'TRACE'
+            logLevel = 'trace';
+            break;
+        case 1: // 'DEBUG'
+            logLevel = 'debug';
+            break;
+        case 2: // 'INFO'
+            logLevel = 'info';
+            break;
+        case 3: // 'WARN'
+            logLevel = 'warn';
+            break;
+        case 4: // 'ERROR'
+            logLevel = 'error';
+            break;
+        case 5: // 'SILENT'
+            logLevel = 'off';
+            break;
+    }
+    const logFn = (level: AvmLogLevel, msg: string) => {
+        switch (level) {
+            case 'error':
+                log.error(msg);
+                break;
+
+            case 'warn':
+                log.warn(msg);
+                break;
+
+            case 'info':
+                log.info(msg);
+                break;
+
+            case 'debug':
+            case 'trace':
+                log.log(msg);
+                break;
+        }
+    };
+    return AirInterpreter.create(handler, peerId, logLevel, logFn);
+};
+
+/**
+ * Checks the network connection by sending a ping-like request to relat node
+ * @param { FluenceClient } peer - The Fluence Client instance.
+ */
+export const checkConnection = async (peer: FluencePeer, ttl?: number): Promise<boolean> => {
+    if (!peer.getConnectionInfo().isConnected) {
+        return false;
+    }
+
+    const msg = Math.random().toString(36).substring(7);
+    const callbackFn = 'checkConnection';
+    const callbackService = '_callback';
+
+    const [request, promise] = new RequestFlowBuilder()
+        .withRawScript(
+            `(seq 
+        (call init_relay ("op" "identity") [msg] result)
+        (call %init_peer_id% ("${callbackService}" "${callbackFn}") [result])
+    )`,
+        )
+        .withTTL(ttl)
+        .withVariables({
+            msg,
+        })
+        .buildAsFetch<[string]>(callbackService, callbackFn);
+
+    await peer.initiateFlow(request);
+
+    try {
+        const [result] = await promise;
+        if (result != msg) {
+            log.warn("unexpected behavior. 'identity' must return the passed arguments.");
+        }
+        return true;
+    } catch (e) {
+        log.error('Error on establishing connection: ', e);
+        return false;
+    }
+};
